Pull up `PYTHONPATH` from `PythonLoader` subclasses refactoring.
